feat(tui): architectural budget test + update.rs split (GH #1018)
New tests/budget.rs: 4 assertions — no src/*.rs over 1000 LOC (app.rs
gets 1300 legacy limit), no async in view.rs/app.rs, size_of<Message> ≤ 128.

update.rs (953 LOC) split into:
- update.rs (637 LOC) — core update fn + key/mouse/modal handling
- update_command.rs (348 LOC) — dispatch_command + extract_selection_from_regions
